pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-borg"
version = "3.3.0"
description = "Collective memory for AI coding agents — your agent learns from every session"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Hermes Team", email = "aleshbrown@gmail.com"},
]
keywords = ["ai", "agents", "workflows", "mcp", "knowledge-exchange", "proof-gates"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "sse-starlette>=1.6.0",
]

[project.optional-dependencies]
crypto = ["pynacl>=1.5.0"]
embeddings = ["sentence-transformers>=2.2.0", "numpy>=1.24.0"]
defi = [
    "aiohttp>=3.9.0",
    "cryptography>=41.0.0",
]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "pytest-asyncio>=0.21.0"]
all = ["agent-borg[crypto,embeddings,defi,dev]"]

[project.scripts]
borg = "borg.cli:main"
borg-mcp = "borg.integrations.mcp_server:main"
borg-http = "borg.integrations.http_server:main"
borg-defi = "borg.defi.cli:main"

[project.urls]
Homepage = "https://github.com/bensargotest-sys/agent-borg"
Repository = "https://github.com/bensargotest-sys/agent-borg"
Documentation = "https://github.com/bensargotest-sys/agent-borg#readme"

[tool.setuptools.packages.find]
include = ["borg*"]
exclude = ["borg.tests*", "borg.defi.tests*", "test_e2e"]

[tool.setuptools.package-data]
borg = ["seeds_data/**/*.md", "seeds_data/**/*.yaml", "seeds_data/**"]

[tool.pytest.ini_options]
testpaths = ["borg/tests", "borg/defi/tests"]
asyncio_mode = "auto"
markers = [
    "network: tests that require network access (borg_search, borg_try, borg_pull against live servers)",
    "defi: tests requiring defi optional dependencies (aiohttp, cryptography)",
    "integration: tests that hit real external APIs",
]